fix(ext2): measure the free space of a directory block from the end of its last record (#360) (#361)
* fix(ext2): measure the free space of a directory block from the end of its last record (#360)

`ext2_append_new_direntry` shrinks the last record of a block to its real
length and puts the new record right after it, but it measured the free space
from the start of that last record rather than from its end. The check
therefore passed with `real_rec_len` bytes less room than it believed, the new
record was placed with too little space, and `ext2_initialize_direntry` wrote
the name past the end of the block buffer, over whatever the slab allocator
kept next to it.

With 4096-byte blocks the suite hit this three times per run: a 6-character
name went 7 bytes past the end of the block, and a 250-character name arriving
behind another one went 255 bytes past it. The name did not reach the disk
either, since only `block_size` bytes of the buffer are written out, so the
record described a name that was not in the block.

The block buffers are one page each, so the bytes written out of bounds land at
offset 0 of the neighbouring slab object. When that object was free, they
replaced its freelist node and the next allocation from that page dereferenced
a filename as a pointer, which is how the bug was found: a page fault in
`list_head_remove` with `cr2 = 0x32302e6a` and the faulting pointer reading
"f.02". When the neighbour was in use instead the write was silent, which it
had been on every run so far.

Measure the space from `it.block_offset + real_rec_len`, and let
`ext2_initialize_direntry` refuse a record that cannot hold its own name so a
recurrence is a reported failure rather than silent corruption. A record that
does not fit now falls through to `ext2_create_new_direntry`, which opens the
next block, exactly as it already did one record later.

`t_dir_block_boundary` covers both shapes of the boundary. It does not fail
against the unfixed code: the name written out of bounds survives in the
neighbouring buffer, so lookups still find it. The deterministic failure of the
unfixed code is the panic above, and the permanent detector is the bound check.

* fix(logging): cast the size arguments the macOS toolchain types differently

The format attributes added in #216, and the `z` coverage added in #332, catch
mismatches that only one of the two cross toolchains reports. `sizeof` and `%z`
are checked against the compiler's `__SIZE_TYPE__`, which is `unsigned int` for
the i686-elf gcc the Ubuntu job uses and `long unsigned int` for the one
Homebrew ships. Both are 32 bits here, so only the type name differs and only
the format check notices. The macOS job has been failing on `develop` since
those landed:

    heap.c:342:17: error: format '%u' expects argument of type 'unsigned int',
    but argument 7 has type 'long unsigned int' [-Werror=format=]

    t_printf_length.c:72:28: error: format '%zx' expects argument of type
    'size_t', but argument 4 has type 'unsigned int' [-Werror=format=]

The `sizeof` arguments are cast to `unsigned`, which matches `%u` on either
toolchain. The `%z` cases cannot be fixed that way, because `%z` is the one
conversion whose expected type is exactly the one the toolchains disagree on:
they use `__SIZE_TYPE__` and `__PTRDIFF_TYPE__`, so each toolchain checks them
against its own spelling. `stddef.h` declares `size_t` as `unsigned long`,
which matches the Homebrew toolchain and not the other one; leaving that alone
here, since changing it would move every format check that involves a `size_t`.

The sites were found two ways, since neither alone is complete: preprocessing
every translation unit of the local build with its own flags, which catches a
`sizeof` that arrives through a macro, and scanning the sources, which catches
a translation unit the local build does not configure.

- `heap.c:342`, where the `sizeof` arrives through the `OVERHEAD` macro.
- `t_printf_length.c`, in the `z` cases of the length-modifier coverage.
- `t_big_write.c:39`, which is built but not in the list `runtests` runs.
- `multiboot.c:162` and `ext2_dir.c`, in `pr_debug` calls that the debug level
  of those files compiles out, so they are latent rather than breaking:
  raising it would break the macOS build too. `multiboot.c` also used `%d` for
  an unsigned count, which is now `%u`.

Four more `%z` sites, in `test_utils.h`, `test_gdt.c`, `test_idt.c` and a
`pr_debug` in `zone_allocator.c`, are not in this build and are left alone.

+++ kernel/src/multiboot.c
@@ -160,8 +160,8 @@ void dump_multiboot(multiboot_info_t *mbi)
     if (bitmask_check(mbi->flags, MULTIBOOT_FLAG_MMAP)) {
         pr_debug("%-16s = 0x%x\n", "mmap_addr", mbi->mmap_addr);
         pr_debug(
-            "%-16s = 0x%x (%d entries)\n", "mmap_length", mbi->mmap_length,
-            mbi->mmap_length / sizeof(multiboot_memory_map_t));
+            "%-16s = 0x%x (%u entries)\n", "mmap_length", mbi->mmap_length,
+            (unsigned)(mbi->mmap_length / sizeof(multiboot_memory_map_t)));
         multiboot_memory_map_t *mmap = mmap_first_entry(mbi);
         do {
             pr_debug(
